face-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
face_recognition==1.3.0
numpy==1.26.4
Pillow==10.3.0
python-multipart==0.0.9
